fix(job-manager): addresses Clément's review comments: move non-generic methods to the derived class SyncJob
