+++ src/libsyncengine/jobs/abstractjob.cpp
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "syncjob.h"
+#include "abstractjob.h"
 #include "log/log.h"
 #include "requests/parameterscache.h"
 
@@ -52,35 +52,6 @@ ExitInfo AbstractJob::runSynchronously() {
     return _exitInfo;
 }
 
-void AbstractJob::setProgress(int64_t newProgress) {
-    _progress = newProgress;
-    if (_progressPercentCallback) {
-        if (_expectedFinishProgress == expectedFinishProgressNotSetValue) {
-            LOG_DEBUG(_logger,
-                      "Could not calculate progress percentage as _expectedFinishProgress is not set by the derived class (but "
-                      "_progressPercentCallback is set by the caller).");
-            _expectedFinishProgress = expectedFinishProgressNotSetValueWarningLogged;
-            _progressPercentCallback(_jobId, 100);
-        } else if (_expectedFinishProgress == expectedFinishProgressNotSetValueWarningLogged) {
-            _progressPercentCallback(_jobId, 100);
-        } else {
-            _progressPercentCallback(_jobId, static_cast<int>((_progress * 100) / _expectedFinishProgress));
-        }
-    }
-}
-
-void AbstractJob::addProgress(int64_t progressToAdd) {
-    setProgress(_progress + progressToAdd);
-}
-
-bool AbstractJob::progressChanged() {
-    if (_progress > _lastProgress) {
-        _lastProgress = _progress;
-        return true;
-    }
-    return false;
-}
-
 void AbstractJob::abort() {
     if (ParametersCache::isExtendedLogEnabled()) {
         LOG_DEBUG(_logger, "Aborting job " << jobId());
@@ -90,13 +61,6 @@ void AbstractJob::abort() {
 
 bool AbstractJob::isAborted() const {
     return _abort;
-}
-
-void AbstractJob::run() {
-    _isRunning = true;
-    runJob();
-    callback(_jobId);
-    // Don't put code after this line as object has been destroyed
 }
 
 void AbstractJob::callback(UniqueId id) {
