+++ src/libsyncengine/jobs/syncjob.cpp
@@ -18,8 +18,45 @@
 
 #include "syncjob.h"
 #include "log/log.h"
-#include "requests/parameterscache.h"
 
 #include <log4cplus/loggingmacros.h>
 
-namespace KDC {} // namespace KDC
+namespace KDC {
+
+void SyncJob::setProgress(int64_t newProgress) {
+    _progress = newProgress;
+    if (_progressPercentCallback) {
+        if (_expectedFinishProgress == expectedFinishProgressNotSetValue) {
+            LOG_DEBUG(_logger,
+                      "Could not calculate progress percentage as _expectedFinishProgress is not set by the derived class (but "
+                      "_progressPercentCallback is set by the caller).");
+            _expectedFinishProgress = expectedFinishProgressNotSetValueWarningLogged;
+            _progressPercentCallback(jobId(), 100);
+        } else if (_expectedFinishProgress == expectedFinishProgressNotSetValueWarningLogged) {
+            _progressPercentCallback(jobId(), 100);
+        } else {
+            _progressPercentCallback(jobId(), static_cast<int>((_progress * 100) / _expectedFinishProgress));
+        }
+    }
+}
+
+void SyncJob::addProgress(int64_t progressToAdd) {
+    setProgress(_progress + progressToAdd);
+}
+
+bool SyncJob::progressChanged() {
+    if (_progress > _lastProgress) {
+        _lastProgress = _progress;
+        return true;
+    }
+    return false;
+}
+
+void SyncJob::run() {
+    _isRunning = true;
+    runJob();
+    callback(jobId());
+    // Don't put code after this line as object has been destroyed
+}
+
+} // namespace KDC
